Move all IXIA-based test cases to ixia folder
